Add bazelisk
We add this in the form of a first-party script to download the tool,
and we check the tool itself into the repo.  We include its license
along with a BUILD file that exports the license.

The script was adapted from an old Aurora script to download bazelisk,
originally written in 2019.