Add `stdx` target (#8)
This code backports library utilities from future versions of the C++
standard to C++14.  This includes "experimental" features, such as
`is_detected`, which lives in an `experimental/` subdirectory.

In cases where we're able to simply copy the reference implementations
from cppreference.com, we do so, and omit unit tests.  However, some
reference implementations rely on C++14-incompatible features, such as
`if constexpr`.  In these cases, we modify the reference
implementations, and add unit tests.