Add make-single-file tool for single-file delivery
Zero-setup usage examples:

```sh
tools/bin/make-single-file au/au.hh > au.hh
tools/bin/make-single-file au/au.hh au/io.hh > au_with_io.hh
```

The first would give a "bare" single-file `au`.  The second would
include our I/O capabilities.

The state of the script is "good enough to start".  One silly bit is
that we will often close the `au` namespace at the end of one file's
contents, only to open it at the start of the next.  We can refine that
later.  We may also consider printing the name of the original source
file as a comment later on, to help users find the original source of
the contents.  But again---this creates a working single file, so it's
"good enough to start".