Add "How-to" guide for defining new units
This is adapted from the AV repo `unit_of_measure` doc.  However, we
wrote that doc before we had a clear understanding of the four
categories of documentation, and it mixed together the "How-to" and
"Explanation" categories.  This new doc is more focused: it's a pure
"How-to" guide, with just enough targeted explanation to clarify the
motivation in a few places.

Technically, we haven't fully landed our documentation system yet.  It's
out for review, in #13.  However, I patched it in and rendered the
website.  The results for this page can be viewed at
(https://aurora-tech.github.io/au/howto/new-units).