Add `au-docs-serve` to make and serve doc site (#13)
Any user who has set up `direnv` can simply run `au-docs-serve`, and
will be presented with a message to view the site on
(http://127.0.0.1:8000/au/).  As doc pages are saved, the site will
regenerate and refresh automatically.  Users can exit with `Ctrl-C`.

All necessary packages are specified inside of `requirements.in`, for
hermeticity and reproducibility.

The `index.md` page is mostly a placeholder for now.  It shows that a
few key features are working as intended.