Add "How-to" guide for defining new units (#16)
This is adapted from the AV repo `unit_of_measure` doc.  However, we
wrote that doc before we had a clear understanding of the four
categories of documentation, and it mixed together the "How-to" and
"Explanation" categories.  This new doc is more focused: it's a pure
"How-to" guide, with just enough targeted explanation to clarify the
motivation in a few places.