Add `--units` parameter to `make-single-file` (#23)
Now that we've split up the units, adding new units in the single-file
script has become extremely tedious, e.g.,

```sh
make-single-file \
  au/au.hh \
  au/io.hh \
  au/units/meters.hh \
  au/units/seconds.hh \
  au/units/grams.hh \
  au/units/kelvins.hh \
  > ~/au.hh
```

We can simplify this quite a bit by making the above equivalent to this:

```sh
make-single-file \
  au/au.hh \
  au/io.hh \
  --units meters seconds grams kelvins \
  > ~/au.hh
```

I tested this change with a downstream PR, coming soon, which will add
unit tests for the single-file versions we generate for the docs.  I
also verified that we can still call the script with no `--units`
supplied.