Add automatic tests for ready-made single files
We will plan to link directly to these files on the "installation" page
of our doc website.  It's important to check, automatically, that
including these files lets us do all the stuff we want to do with the Au
library.

To achieve this, we create a `cc_library` to go with each genrule.  We
then add a `cc_test` rule for each of these libraries to put them
through their paces.  Common test cases are factored out into a separate
file for maintainability.

What do we "expect" from our single-file installations?

1. Basic Quantity/QuantityPoint functionality.
2. Prefixes.
3. Math functions.
4. Base units for each dimension, including dimensionless (Unos).

By creating these tests, we uncovered that several of these pieces of
functionality were missing.  We weren't previously including `prefix.hh`
(and this PR fixes that).  We also weren't including most of the units,
ever since #17 landed and split them into their own files.