Include au.hh and io.hh by default in single file
We should take advantage of the fact that this script is only for our
repo, and not for general use, by making the default use cases as simple
as possible.  After this PR, end user installation will look like one of
these possibilities:

```sh
make-single-file --units meters seconds grams kelvins > ~/au.hh

make-single-file --noio --units meters seconds grams kelvins > ~/au.hh
```

The `--prefix` parameter wasn't ever really changed.  It's a holdover
from the proof of concept version of this script from the AV repo.  In
its place, we have `--noio`.  Including `<iostream>` support is a good
default, and this option makes it easy to opt out of that default.

I tested this by running the tests in #24, after patching this in and
fixing up the commands.