Add remaining SI base units (#25)
To do this, we needed to add new base dimensions for `AmountOfSubstance`
and `LuminousIntensity`.

Each unit needs a corresponding test file which verifies not only the
unit symbol, but also whatever relationships with other units we need to
be confident that it is defined correctly.  For this reason, we also
created a few auxiliary units:

- `katals` (equivalent to `moles / second`)
- `lumens` (equivalent to `candela * steradians`)

These relationships can be found on this page:
https://www.nist.gov/pml/owm/metric-si/si-units

The latter one required us to (finally) define the steradian, which is
equivalent to the squared radian.

Since these are base units, we also include them in the single-file
releases automatically.

Fixes #18.