Unlock value member handle with equivalent unit
We add a `.data_in(u)` member for `Quantity` and `QuantityPoint`.  This
gives a direct handle to the value member of the underlying numeric type
(Rep).  To "unlock" it, `u` must be an instance of any _equivalent_ Unit
or Q-Maker.  Explicitly:

- For `Quantity`, we use quantity-equivalence, and the "Q-Maker" is
  `QuantityMaker`.
- For `QuantityPoint`, we use point-equivalence, and the "Q-Maker" is
  `QuantityPointMaker`.

In case this sounds too complicated, these examples should clarify.

- If `q = feet(3)`, then `q.data_in(feet)` or `q.data_in(Feet{})` will
  _work_, because we provided the same unit.
- If `q = kilo(hertz)(5)`, then `q.data_in(inverse(milli(seconds)))` or
  `q.data_in(UnitInverseT<Milli<Seconds>>{})` will _work_, because `q`
  is a `Quantity` and these units are _quantity-equivalent_.
- If `p = celsius_pt(20)`, then `p.data_in(celsius_pt)` or
  `p.data_in(Celsius{})` will _work_.  `p.data_in(kelvins_pt)` will
  _not_ work, because `p` is a `QuantityPoint` and Kelvins are _not
  point-equivalent_ to Celsius.

For mutable handles, we test the functions by... well, mutating.  For
const handles, we need to test something about the addresses of the
object, and of its `.data_in()` result.  I tried writing a reasonable
utility, and gave it a few tests of its own.

Unit-based overloads are implemented _in terms of_ Q-maker based
overloads, and not vice versa, because the latter are by far more
common, and we want the commoner case to be faster to compile.

Fixes #26.

Test plan:
- [x] Include unit tests
- [ ] Measure compile time impact on benchmarks and real targets