Add installation instructions
I tried to make it easy to understand the different options we provide,
including a coloured comparison table and a flowchart.  I then included
detailed instructions for each option.

The table colour scheme is chosen to be colourblind-friendly.  I added
some custom CSS to achieve this.  I also enabled mermaid support, which
made it easy to make the flowchart.

There are some aspects still missing.  We can't _fully_ test the bazel
installation until the library is publicly available.  There are also a
bunch of doc pages that don't yet exist, but which we'd like to link to
before release.  However, overall, it should be good enough to land as a
first version.