Print manifest comment in single-file packages (#36)
Once `au` goes public, there are going to be a _lot_ of files out in the
wild called `au.hh` (and a bunch more called `au_noio.hh`).  These files
are all going to look very similar when people skim them, but they'll
have myriad tiny variations.  It's critically important to make these as
easy as possible to distinguish!

To nip this in the bud, we'll add a comment at the top of the file
which provides a "manifest", clearly identifying both the version of
code and the options with which the file was generated.  We'll add a new
`--version-id` argument to the script, which gives a human-readable,
release-tag-based description.

Right this moment, it gives: `0.2.0-1-g545a831-dirty`.  This means we
are based on release `0.2.0`, with `1` extra commit, and the current
unambiguous ID is a git (`g`) commit with abbreviated hash `545a831`.
Since the worktree is dirty (after all, I'm in the _process_ of
committing), it appends the word `dirty`.  If we were exactly on a
release, the commit hash would be omitted.  Since we don't want to
burden end users with generating this information, we will provide it if
they omit it.

Git version information is typically inaccessible to bazel.  This makes
it hard to include the manifest in our automatically generated
single-file releases.  To fix this, we use bazel's workspace status
machinery [1], writing the ID to the `stable-status.txt` file.  We then
give the genrules access to this file.

To do so, we need to set the `--workspace_status_command` for every
build command in our `.bazelrc`.  Typically, one would rather not do
this for _every_ command.  However, empirically, the burden was too
small to be perceptible.

Fixes #30.

Test Plan
---------

- [x] Run `au-docs-serve`, and view the single-file releases.
- [x] Run `make-single-file --units meters seconds | gview -`.

[1] https://bazel.build/docs/user-manual#workspace-status