Migrate `aurora-tech` to `aurora-opensource` (#44)
Now that the repository has been moved to its final home, many of the
URLs are out of date.  Most of these are in the unit definition files.

Among the rest, the most interesting is in the installation doc.  We
needed to be sharp.  There was some ASCII art to realign.  More
seriously, the SHA256 sum for the release tarball changed after we moved
the repository, so I updated it to make sure the instructions were fully
up to date.