Automatically enforce copyright header
This adds a github workflow to check the license header automatically
on almost every file in the repo.  We have used the
viperproject/check-license-header check, and pinned it to the specific
SHA corresponding to v1.0.5 for security purposes.

To the extent that we find any files that are out of compliance, we will
fix them in this PR.