Automatically enforce copyright header (#47)
This adds a github workflow to check the license header automatically
on almost every file in the repo.  We have used the
viperproject/check-license-header check, and pinned it to the specific
SHA corresponding to v1.0.5 for security purposes.

To the extent that we find any files that are out of compliance, we will
fix them in this PR.

Note that any files not covered by any existing rule will be flagged as
uncovered, and we have set this up so that it fails the build.  Thus,
this will handle new files going forward as well as existing ones.