Only rebuild for this present branch
The `git describe` is working again, but it's referring to weird merge
commits such as 7f283f0c10ff6adf333a8e207001ed7812a1a2a2.  I can't
figure out what is creating these commits.

One of the parents happens to be the current HEAD of main.  This makes
it seem likely that it would "just work" if we restrict it to main
(which we want to do anyway), because the attempt to merge that branch
with itself should be a no-op.  I'm going to try triggering on this
present branch first though.