Add "Au 101" tutorial: quantity makers
I think the most basic simple chunk to introduce people to the library
is to have them make some quantities, using the quantity makers.  I even
deferred the `Quantity` type names for a future lesson.

The basic structure is as follows:

- https://aurora-opensource.github.io/au/tutorial/101-quantity-makers
  The main tutorial page, which includes explanatory text.

- https://aurora-opensource.github.io/au/tutorial/exercise/101-quantity-makers
  An interactive exercise, including revealable example solutions.

- `tutorial/101_quantity_makers.cc`
  The file which users will edit in the course of completing the
  exercise.

I've left a TODO to link to the development page, because we haven't
added that doc yet.  This is part of a larger strategy for getting the
docs done sooner, without wasting time trying to put the PRs out in
topologically sorted order.  It's easy to grep for `TODO`, and we can do
a `TODO`-cleanup pass on the docs right before we flip the switch.

Reviewers should go through the tutorial to see how it works.
Instructions:

1. `bonsai grab --parents 55`
2. `au-docs-serve`
3. Load up the tutorial webpage at the URL that will be printed from
   that command, and follow the instructions.  (You may want a separate
   terminal window once you start running the bazel commands.)

Note that the not-yet-existent development page will include
instructions for setting up `direnv`, and will explain that we assume
users have done this.  If you have not done this, and don't want to,
then you'll need to prepend `tools/bin/` to all of the commands (such as
`bazel` and `au-docs-serve`).