Add "Au 102" tutorial: API types
This builds on the previous example by explaining how to spell the
quantity type names, and showing how we can use them to make unit-safe
interfaces.

As with #55, reviewers should go through the tutorial to check it for
flow and content.

All other considerations (i.e., TODOs, URLs, instructions) are the same
as in #55.