Add buildifier 5.1.0 (#50)
This PR creates `tools/bin/buildifier`, so we can all use a known
version of buildifier.  I used the same strategy we used for
`tools/bin/clang-format`: a thin wrapper around `bazel run`, with
bazel's output suppressed to make it feel like a command.

This also suffers from the same drawback: the first run is really slow.
I filed #49 to track this.

Unfortunately, `--version` does not work; it shows up as "redacted".
This is being tracked in bazelbuild/buildtools#831.