Configure license checking on GitHub workflow files (#63)
### Summary

I noticed we aren't enforcing license checking on any of our GitHub
workflow files. The reason is that the generic glob pattern we're using
(`**/*.yml`) doesn't match anything in hidden directories. I submitted a
feature request
[here](https://github.com/viperproject/check-license-header/issues/350)
to make this more flexible. In the meantime, explicitly list out
`.github` in the config.

### Test Plan
* update the config first, see the license check
[fail](https://github.com/aurora-opensource/au/actions/runs/3857899619/jobs/6575810317)
for `clang-format-lint`