Add developer's guide
This page explains how users can set up Au for development.

We list the two main use cases: library contributions, and tutorial
exercises.  Then we list the core steps which both use cases have in
common.  Finally, we explain how to perform the most common types of
workflows: building and testing code, and rendering and serving the
documentation website.

Any admonitions which are only for subsets of users are made
collapsible, to promote the readability and flow of the webpage.