Reorder rows in very rough importance order
First two rows: can I use it _at all_, and if so, how easy is it to get?

Next two: will it kill my build times, or build error interpretability?

Then we get into the library features: unit safety, low friction,
composability, mixed-rep support, etc.

The order is very rough, but it's probably good that units-as-types, and
macros, are at the bottom.