Add GitHub Action workflow ownership via CODEOWNERS (#65)
The `CODEOWNERS`
[file](https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/about-code-owners)
is used to track ownership throughout a repo. Set it up to add the
"GitHub Actions Reviewers" team on any PRs that modify workflow files.

I'm unsure how to test this until after landing on mainline, so I'll
keep an eye on future workflow changes (or put up a test PR afterwards).
EDIT: GitHub has native validation of the CODEOWNERS file in the "Files
changed" view.