Expand copyright notice for single-file packagings (#67)
Files in our repo use a one-line copyright notice, which is sufficient
because they will always be distributed alongside the `LICENSE.txt`
file.

The generated single-file versions of the library are different: they
are intended to be packaged alone.  There will be copies of variants of
these files all over all kinds of machines.  Thus, it makes sense to
include the full copyright notice for these files.

To test this, run `au-docs-serve`, and observe the following header in
the generated files:

```cpp
// Copyright 2023 Aurora Operations, Inc.
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//    http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.
```