Add troubleshooting page (#59)
Whenever anyone has a compiler error, this page will be the one-stop
shop to understand what it means, and how to solve it.

We've included a specific file in the repo which contains an example of
each of these errors. The text of the compiler errors comes from the
compiler versions we officially support. We also include a tabbed
section which has one tab for each snippet of broken code, and another
tab showing how to fix it.