Prepare "entry" docs for public consumption
On the README.md front, this means providing the "quick sell" for the
library: this will be the first thing most users read when stumbling
across the repository.  I tried to write enough to give a clear
impression, with links to more detail, without being exhaustive.

On the documentation website main page, I cleaned up the placeholders,
and gave an introduction and a guide to a few pages.  For some of these
pages, I found myself wanting to link to an "index" page for a section,
so I added that index where appropriate (i.e., tutorials and how-to).

In both cases, we added the appropriate logos, including the favicon for
the documentation website.