Add page comparing Au to alternatives (#57)
The first question many will ask is: why another C++ units library?
We'll want to have a crisp, easily accessible answer.

This page provides that answer, in two sections.

First, we give the big picture.  Au addresses the most common
"dealbreakers" that prevent people from using the existing alternatives.
Moreover, we have a variety of either rare or outright novel features.

Second, we get into the weeds.  We provide 20 criteria, and assess both
Au and the 3 most prominent alternatives on (almost) all of them.  (Some
assessments we've deferred for now, simply because this is a _lot_ of
work.)  We also link directly to a fuller explanation of what each
criterion means.

Later on, we may follow up by providing a comparison page specific to
each major alternative units library, living under the `alternatives/`
folder in the documentation website.  But this page is the most
important to get done before public release.