Remove TODO warning in forking instructions
This PR was created from a clone of the main repository, which was
transitioned to a personal fork which was created after the fact.  I
followed the instructions in the development guide.  Therefore, if this
PR works, then we should be confident enough in these Instructions to
remove the TODO warning.