Add core reference docs
This PR creates both the `reference` section and the `discussion`
section: two of the four major categories of documentation.  Since these
pages tend to reference each other a lot, I decided to combine them in a
single PR which we could use as foundational.

The vector space doc is migrated from the version in Aurora's repo,
almost unmodified.

In the reference section, I wanted to make sure to hit Dimension,
Magnitude, and Unit.  All three of these types share a common property:
namely, that we often create instances of them, and that there is only
one possible value for the instances, so that types and values are
effectively completely interchangeable representations of the same
"thing".  I decided to invent a name, "monovalue types", to describe
this concept, because it turns out to be foundational to Au.  Speaking
of foundational, I migrated the parameter packs documentation from
Aurora's repo as well.

I updated the installation instructions to replace the "before public
release" language with a `TODO` that will age better.

The last three subsections of the unit doc are also `TODO`, because I
ran out of steam.