Keep QuantityPoint conversion within its Rep (#101)
If we call `celsius_pt(x).in(kelvins_pt / mag<20>())`, the result should
have the same type as `x`.  Currently, that's not the case for two
reasons, both of which stem from the underlying operation of
_subtracting the origin displacement_.

1. The origin itself typically has `int` Rep.

2. Even if it didn't, C++'s integer promotion rules would take us
   outside the realm of shorter integral types.

Part of the solution is to cast the origin displacement Rep to the
desired Rep type explicitly.  There's an opportunity for a safety check
here, so we add it (including unit tests).

The rest of the solution is simply to `rep_cast` at the end to "undo"
any integer promotion that may have happened.

Along the way, we also changed the operation from _subtraction_ of one
origin displacement, to _addition_ of the inverse displacement.  Now
that we're being more careful about checking whether the origin
displacement fits in the Rep, this would have broken unsigned type
support if we hadn't made this change.  (What I think was happening
before with the unsigned types was probably promotion to signed integers
--- so, no "incorrect" results, but probably an unexpected Rep.)

We also needed to add a trivial `rep_cast` overload for `Zero`.  This
helps it "stand in" for arbitrary quantities in one more way.

Fixes #100.

Test plan
---------

- [x] Added new test cases for `OriginDisplacementFitsIn`.
- [x] Uncommented the test case broken in #100, and verified that it
      fails for the right reason.
- [x] Test that there is no unacceptable compile time penalty.