Tweak overflow and inverse thresholds (#102)
The primary motivation is to move to a more defensible policy for
inversions.  Formerly, our policy was "avoid truncating nonzero values
to 0" for any value <= 1000.  The new policy is "keep round-trip
inversion lossless" for any such value.  This amounts to changing the
threshold from 1,000 to 1,000,000.

Unfortunately, if we made this change alone, we'd have poor ergonomics
even for pretty reasonable conversions.  For example:

```cpp
constexpr auto period = inverse_as(micro(seconds), hertz(40));
```

This fails because converting `1` to the (dimensionless) product of the
units multiplies by 1,000,000.  Since we're using `int`, this would
overflow the old threshold of 4,294.  So, these thresholds are
connected.

I've actually been thinking for a while that it would be nice to loosen
that threshold ever so slightly.  It's a nuisance to treat `int32_t` and
`uint32_t` differently, despite a very small difference in their ranges.
So I also make that change in this PR.  Ultimately, I think this will be
a very durable "sweet spot".  It's pleasing that we prioritize values of
up to 1,000 generally, and you can fit 3 factors of 1,000 in an `int`,
with a little room to spare (which we put towards extending the safety
surface).

Fixes #69.

Test plan
---------

- [x] Existing unit tests (subject to appropriate tweaks).
- [x] Tested on Aurora's whole codebase.  No significant errors
      introduced.
- [x] Checked compile time impact: no significant effect.
- [x] Added a test, `QuantityPoint.PreservesRep`, which I had wanted to
      add in #101, but couldn't.