Port XKCD 2585 test to Au repo
A lot has happened since we first reproduced XKCD 2585[1] in our units
library[2].  At the time, the library where we added the test was just
"Aurora's units library".  The library Au, which started out as a
drop-in replacement for "Aurora's units library" and has now been open
sourced, wouldn't even be conceived of for another month or so.

Now that Au is out, it's time to reimagine this unit test to take
advantage of its more fluent, natural interfaces.  Instead of clunky
`quotient_t` metafunctions and template parameters, we have simple
division and function parameters.  We can also form compound units on
the fly by composition --- for example, `kilo(meters) / hour` instead of
`KilometersPerHour`.

This results in the addition of four more units to the library, so of
course they get their own (rather less whimsical) unit tests, too.

[1] https://xkcd.com/2585/
[2] https://mobile.twitter.com/openMindedSkep/status/1496964956390137858

Test plan
---------

I added printing code to this implementation, so we could check the
intermediate results, too --- taking advantage of Au's automatic unit
label generation facilities.