Ensure that `Rep` is `std::is_arithmetic` (#104)
This applies to both `Quantity` and `QuantityPoint`.  Most users tend
not to hit this use case, but when they do, it can be pretty confusing.
This `static_assert` will give a direct, readable error message.

It turns out we really only need/want to add this to `Quantity`, not
`QuantityPoint`. If we add it to both, we get a double error message for
`QuantityPoint`.

Fixes #38.

Test plan
---------

I tried some silly test cases, like `meters(meters(3))`, and
`meters_pt(meters_pt(3))`.  They fail to compile both with and without
this PR in both cases, but the error message with this PR is _much_
shorter and more direct.