Compile with  `-Wextra` (#114)
This patch adds `-Wextra` to the compiler options for the three
compilers we currently support. It's a bit uglier than I'd hoped
because we can't specify extra options. We can only override the list
of options. So we have to enumerate the full list manually in clang's
case. For GCC, we have access to the default list.
    
I validated manually with `bazel aquery` that the only difference
before and after this patch is the added `-Wextra` flag.