Add operations for QuantityMaker and SingularNameFor
We should let users multiply or divide two `QuantityMaker` instances if
they want to.  It won't usually be gramatically correct, but it probably
doesn't rise to the level of an "error".  It's also necessary for units
whose singular name equals their plural name, such as `hertz`.  This
fixes up an inconsistency in the docs, where we called the
`SingularNameFor` "optional", although previously it really wasn't.

We also need to permit multiplying `SingularNameFor` instances.  This
was always simply an oversight, preventing legitimate use cases such as
`(radians / (meter * second))`.

Fixes #72.

Fixes #71, since that task was to document an error in the
troubleshooting guide, but this PR removes that error.