Add operations for QuantityMaker and SingularNameFor (#118)
We should let users multiply or divide two `QuantityMaker` instances if
they want to.  It won't usually be gramatically correct, but it probably
doesn't rise to the level of an "error" (which is how we currently treat
it).  It's also necessary for units whose singular name equals their
plural name, such as `hertz`.  This fixes up an inconsistency in the
docs, where we called the `SingularNameFor` "optional", although
previously it really wasn't.

We also need to permit multiplying `SingularNameFor` instances.  This
was always simply an oversight, preventing legitimate use cases such as
`(radians / (meter * second))`.

I investigated the compile time impacts, and confirmed that any possible
impact is well within the measurement noise.

Fixes #72.

Fixes #71, too, since that task was to document an error in the
troubleshooting guide, but this PR removes that error.

Test Plan
---------

- [x] Add new unit tests.
- [x] Investigate compile time impacts.
- [x] Pass PR builds.