Upgrade certifi and other python requirements (#126)
The certifi update fixes a security issue:
https://github.com/aurora-opensource/au/security/dependabot/1

As for the other updates, I suppose we're due for a refresh.

Test Plan
---------

I rendered the doc website and spot-checked the changes.  It looks like
with the new mkdocs version, we need to "opt in" to the footer links
that we used to get by default, so I added the relevant option.