Add `//:mike` command for versioned docs
`mike` (https://github.com/jimporter/mike) is a tool for managing
versioned documentation when using both mkdocs and GitHub pages (which
we are).  The idea is that you render a given version of the docs once,
store it in your `gh-pages` branch, and never touch it again.  `mike`
will store it in a subfolder with that version's name.  You can also set
up a combo box on each page, so users can switch to see the current page
at any version of the docs.

This PR is just the first step for #131.  We'll add a new target,
`//:mike`, which exposes the tool: to run,

```sh
mike ...
```

you would instead run,

```sh
bazel run //:mike -- ...
```

To prove that it works, we'll cherry-pick this commit onto `0.3.1` and
`0.3.2` and generate versions of the documentation website.  (We won't
do `0.3.0` or earlier, because those versions predate public
availability, and they had pretty embarrasing placeholder content.)
We'll publish these doc site versions to `gh-pages`, so they'll be
publicly available, but nothing will link to them yet.  We'll also
publish a version for `main`.

We also update the instructions for cutting releases to take versioning
into account.

Future steps to resolve the "core" of #131:

1. Make another PR so the GitHub action that updates the doc website on
   every commit to `main` uses `mike`.

2. Set `main` as the "default" in `mike`, which should set up a redirect
   for the root website.  (This won't be a PR; we'll commit directly to
   `gh-pages`.)

3. Replace every existing HTML page on the current website with a
   redirect to `main`, patterned off of the main page.  (This won't be a
   PR; we'll commit directly to `gh-pages`.)