Replace generated files with correct versions
In order to generate the doc site for old versions of the library, which
predate the `mike` tool, we needed to cherry-pick the commit that added
`mike`.  Unfortunately, this changed the git commit description, from,
say, `0.3.2` (for a "pure" checkout) to `0.3.2-1-g1d5079e`, indicating
that there was one (`-1-`) extra commmit after the tag.

To fix this, I checked out each tag and generated authoritative
versions.  Then I copied them into the `gh-pages` branch.  Of course,
since the `mike` commit doesn't otherwise affect the contents of the
files, the net effect is simply to trim the chaff from the end of the
version identifier.  But the files really were generated in the
authoritatively correct way --- I promise!

This will not be a problem for future commits, because `mike` will
already be installed.