Add `//:mike` command for versioned docs (#132)
`mike` (https://github.com/jimporter/mike) is a tool for managing
versioned documentation when using both mkdocs and GitHub pages (as we
do).  The idea is that you render a given version of the docs once,
store it in your `gh-pages` branch, and never touch it again.  `mike`
will store it in a subfolder with that version's name.  You can also set
up a combo box on each page, so users can switch to see the current page
at any version of the docs.

This PR is just the first step for #131.  We'll add a new target,
`//:mike`, which exposes the tool: to run,

```sh
mike ...
```

you would instead run,

```sh
bazel run //:mike -- ...
```

To prove that it works, we'll cherry-pick this commit onto `0.3.1` and
`0.3.2` and generate versions of the documentation website.  (We won't
do `0.3.0` or earlier, because those versions predate public
availability, and they had pretty embarrasing placeholder content.)
We'll publish these doc site versions to `gh-pages`, so they'll be
publicly available, but nothing will link to them yet.  We'll also
publish a version for `main`.

We also update the instructions for cutting releases to take versioning
into account.

Finally, we change the auto-deploy GitHub action so that it uses `mike`
instead, auto-updating the `main` version.  We want to stop clobbering
the existing URLs, because we'll need to manually replace them with
redirects to the corresponding versioned pages.