Manually redirect every non-index page to `0.3.2/`
The index page should always go to the main page as the site evolves.
But the other pages that happen to exist now are different.  They might
move to new locations.

We want anyone with an old URL hanging around to be redirected to that
exact version of the page, which is 0.3.2, because we know that it
exists.