Add most missing core docs
This is the biggest step towards getting the documentation website
"core-complete".  By this I mean that we might be missing tutorials or
how-to guides that I'd _like_ to write, _someday_, but we shouldn't have
any stubs or outright missing pages for the core parts of the library.
After this PR, the main "core" doc still remaining will be the
`QuantityPoint` reference page.

I started by adapting a treasure trove of Aurora-internal Au docs, which
I didn't know what to do with back when we open-sourced Au.  These
include the "concepts" discussion pages `arithmetic.md`,
`common_unit.md`, and `dimensionless.md`.  I rounded this section out
with new discussion docs on `quantity_point` and `zero`.

I also added a new "Au idioms" discussion section, which gave me a home
to migrate the explanation of "unit expression APIs".  This is very
useful and many pages link to it.

The reference docs have been greatly fleshed out.  I finally finished
the stubs for `prefix`, `quantity`, and `unit`.  I also added new
reference docs for `corresponding_quantity`, `math`, and `zero`.

Some of the deeply-nested reference pages --- most especially `math` ---
became hard to read: the `<h4>` sections blended together.  I therefore
added a script which would add a `<hr>` at the end of any `<h4>`
section.  I found this improves readability significantly where it's
used.  We need to explicitly opt into using this on a per-page basis,
which is also fine with me.  (The code is executed directly when the
script loads, rather than being encapsulated in a function, because I
couldn't find a simple way to reliably get the script to finish loading
before I called the function.  This bothers me only very slightly.)

Some of the "main section pages" --- namely, how-to, and reference ---
looked a little sparse.  I decided to add descriptions of the most
significant pages in that section (or, for some, all of the pages).

Finally, there are a couple of code updates.

- Reworded the phrasing in some comments.
- Added a couple test cases to reflect some of the new doc content.
- Finally explicitly deleted the `Zero` constructor for `QuantityPoint`